Rename `rustc_privacy/src/errors.rs` into `rustc_privacy/src/diagnostics.rs`
