+++ compiler/rustc_privacy/src/diagnostics.rs
